Midterm demo

+++ app/src/main/java/com/hcmut/test/geometry/equation/LineEquation.java
@@ -8,7 +8,7 @@ import com.hcmut.test.geometry.Vector;
 public class LineEquation {
     // ax + by + c = 0
     private final float a, b, c;
-    private static final float EPSILON = 1e-6f;
+    private static final double EPSILON = 1e-6;
 
     public LineEquation(float a, float b, float c) {
         this.a = a;
